zero char*ish arguments passed by pointer early
in some error paths, we were not zeroing pointer arguments that
were passed by reference. This was error prone: if a caller failed
to check the return value then they could end up using whatever
was on the heap.

This lead to one possibly-exploitable bug in ssh_rsa_verify()

Spotted by Ben Hawkes